renderings complete

+++ src/ui/renderer.rs
@@ -4,14 +4,14 @@ use eframe::egui;
 pub fn draw_nest(
     ui: &mut egui::Ui,
     world: &World,
-    nest_position_x: usize,
-    nest_position_y: usize,
+    nest_position_x: f32,
+    nest_position_y: f32,
     nest_texture: &Option<egui::TextureHandle>,
     available_size: &egui::Vec2,
     painter: &egui::Painter,
 ) -> () {
-    let ui_x = nest_position_x as f32 * (available_size.x / world.width as f32) - 10.0;
-    let ui_y = nest_position_y as f32 * (available_size.y / world.height as f32) - 10.0;
+    let ui_x = nest_position_x * (available_size.x / world.width as f32) - 10.0;
+    let ui_y = nest_position_y * (available_size.y / world.height as f32) - 10.0;
 
     let nest_icon_size_x = available_size.x / 30.0;
     let nest_icon_size_y = available_size.y / 30.0;
@@ -34,8 +34,8 @@ pub fn draw_nest(
 pub fn draw_ant(
     ui: &mut egui::Ui,
     world: &World,
-    x: usize,
-    y: usize,
+    x: f32,
+    y: f32,
     ant_texture: &Option<egui::TextureHandle>,
     available_size: &egui::Vec2,
     painter: &egui::Painter,
@@ -44,8 +44,8 @@ pub fn draw_ant(
     let ant_icon_size_y = available_size.y / 60.0;
 
     // -5.0 to center the ant's UI body.
-    let ui_x = x as f32 * (available_size.x / world.width as f32) - 5.0;
-    let ui_y = y as f32 * (available_size.y / world.height as f32) - 5.0;
+    let ui_x = x * (available_size.x / world.width as f32) - 5.0;
+    let ui_y = y * (available_size.y / world.height as f32) - 5.0;
 
     let ui_pos = egui::Pos2::new(ui_x, ui_y);
